Start to create a Climb Boy editor which base on KK. Mainly in order to integrate tile map editor and behaviors module, strengthen the visual operation.

+++ ClimbBoy/ClimbBoy/GameClasses/Behaviors/PlaceItemBehavior.h
@@ -21,7 +21,7 @@
 }
 
 @property (nonatomic, weak) SKNode *item;
-@property (nonatomic) int trackedTouch;
+@property (nonatomic) NSUInteger trackedTouch;
 @property (nonatomic) CGFloat placeRange;
 @property (nonatomic, readonly) CGPoint placePoint;
 @property (nonatomic, readonly) CGVector placePointNormal;
